Run tests in CI and added a unit test
